quran-stream-bot/requirements.txt
discord.py[voice]>=2.0.0
aiohttp>=3.8.0
mutagen>=1.45.1
